Allowed ability to place multiple objects
Can now place multiple objects on the scene, and can properly choose what object is placed, but the script used for transforming them affected every at once, working on making it selectable, as of now user can only place the objects but cant manipulate.

Using Lean Touch Scripts for this.