Added the prefabs, readded the place button
I completed all prefabs/textures for the earth layers, and re-added the place button for the geology section of the app, issues arose with the draw tool so needed to work it to start working properly, likely something went wrong with importing the earth asset as at that point the project added things such as vulkan renderer which was not compatible. Everything should be good now, was able to build onto my phone and the drawing works again. Will need to adjust the place button so that all the layers will spawn at where the user is looking or at least the size is large enough to be easily found